s3 integration

+++ FileHandlers/LoalStorageFileHandler.js
@@ -20,44 +20,50 @@ class LocalStorageFileHandler {
     }
 
 
-handleNewFile = async (file, allowedExtensions, finalFolder) => {
-    const fileExtension = path.extname(file?.originalname).toLowerCase().substring(1);
+    handleNewFile = async (file, allowedExtensions, finalFolder) => {
+        const fileExtension = path.extname(file?.originalname).toLowerCase().substring(1);
 
-    if (!allowedExtensions.includes(fileExtension)) {
-        throw { message: "Invalid image file extension", status: 400 };
+        if (!allowedExtensions.includes(fileExtension)) {
+            throw { message: "Invalid image file extension", status: 400 };
+        }
+
+        let folder = path.join(this.basePath, finalFolder,  new Date().toISOString().split('T')[0]);
+
+        if (!fs.existsSync(folder)) {
+            fs.mkdirSync(folder, { recursive: true });
+        }
+
+        let number = 0;
+        let filename
+        let additive = ''
+
+        do {
+            filename = `${path.parse(file?.originalname).name.replace(/ /g, '_').replace(/[#?&]/g, '')}${additive}.${fileExtension}`;
+            additive = `(${number})`
+            number++;
+        } while (fs.existsSync(path.join(folder, filename)));
+
+        folder = path.join(folder, filename)
+        fs.writeFileSync(folder, file.buffer, "binary");
+        return folder
     }
 
-    let folder = path.join(this.basePath, finalFolder,  new Date().toISOString().split('T')[0]);
+    handleReplaceFile = async (file, allowedExtensions, filepath, finalFolder) => {
+        const newFile = this.handleNewFile(file, allowedExtensions, finalFolder)
 
-    if (!fs.existsSync(folder)) {
-        fs.mkdirSync(folder, { recursive: true });
+        //removes old file
+        if (newFile && filepath) {
+            if(fs.existsSync(filepath)){
+                fs.unlink(filepath, (err) => {if(err) throw err})
+            }
+        }
+
+        return newFile
+
     }
 
-    let number = 0;
-    let filename
-    let additive = ''
-
-    do {
-        filename = `${path.parse(file?.originalname).name.replace(/ /g, '_').replace(/[#?&]/g, '')}${additive}.${fileExtension}`;
-        additive = `(${number})`
-        number++;
-    } while (fs.existsSync(path.join(folder, filename)));
-
-    folder = path.join(folder, filename)
-    fs.writeFileSync(folder, file.buffer, "binary");
-    return folder
-}
-
-handleReplaceFile = async (file, allowedExtensions, filepath, finalFolder) => {
-    const newFile = this.handleNewFile(file, allowedExtensions, finalFolder)
-
-    //removes old file
-    if (newFile && filepath) {
-        fs.unlink(filepath, (err) => {if(err) throw err})
-    }
-
-    return newFile
-
+    async processFileLocation(file) {
+        return `${process.env.HOST}:${process.env.PORT}/api/${file}`
     }
 }
